requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.10.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
requests==2.28.1
waitress==2.1.2
Pillow==9.3.0
email-validator==1.3.0

# Development tools
pytest==7.4.0
black==23.3.0
flake8==6.1.0
